Video: refactor Circle/Eclipse drawing
add PlotCircle and PlotEllipse to Geometry.h/cpp
implement Video drawing by delegating to the Geometry functions

drop DrawEllipseSegment, it is slow/complex and only used for reticles
we can plot an ellipse and divide it into segments in GameControl::DrawTargetReticle

+++ gemrb/core/GUI/GameControl.h
@@ -167,7 +167,7 @@ public:
 
 	// GameControl always needs to redraw unless we arent in a game (disabled)
 	bool IsAnimated() const override { return !IsDisabled(); }
-	void DrawTargetReticle(int size, const Color& color, const Point& p) const;
+	void DrawTargetReticle(uint16_t size, const Color& color, const Point& p) const;
 	/** Draws the target reticle for Actor movement. */
 	void DrawTargetReticle(const Movable* target, const Point& point) const;
 	/** Sets multiple quicksaves flag*/
